+++ runtime/cudaq/dynamics/operator_sum.cpp
@@ -501,7 +501,7 @@ std::string sum_op<HandlerTy>::to_string() const {
 
 template <typename HandlerTy>
 complex_matrix sum_op<HandlerTy>::to_matrix(
-    std::unordered_map<std::size_t, int> dimensions,
+    std::unordered_map<std::size_t, int64_t> dimensions,
     const std::unordered_map<std::string, std::complex<double>> &parameters,
     bool application_order) const {
   auto evaluated =
@@ -522,7 +522,7 @@ complex_matrix sum_op<HandlerTy>::to_matrix(
 
 template <>
 complex_matrix sum_op<spin_handler>::to_matrix(
-    std::unordered_map<std::size_t, int> dimensions,
+    std::unordered_map<std::size_t, int64_t> dimensions,
     const std::unordered_map<std::string, std::complex<double>> &parameters,
     bool application_order) const {
   auto evaluated = this->evaluate(
@@ -547,7 +547,7 @@ complex_matrix sum_op<spin_handler>::to_matrix(
   template std::string sum_op<HandlerTy>::to_string() const;                   \
                                                                                \
   template complex_matrix sum_op<HandlerTy>::to_matrix(                        \
-      std::unordered_map<std::size_t, int> dimensions,                         \
+      std::unordered_map<std::size_t, int64_t> dimensions,                     \
       const std::unordered_map<std::string, std::complex<double>> &params,     \
       bool application_order) const;
 
@@ -1721,7 +1721,7 @@ sum_op<HandlerTy> sum_op<HandlerTy>::random(std::size_t nQubits, std::size_t nTe
 
 HANDLER_SPECIFIC_TEMPLATE_DEFINITION(spin_handler)
 csr_spmatrix sum_op<HandlerTy>::to_sparse_matrix(
-    std::unordered_map<std::size_t, int> dimensions,
+    std::unordered_map<std::size_t, int64_t> dimensions,
     const std::unordered_map<std::string, std::complex<double>> &parameters,
     bool application_order) const {
   auto evaluated = this->evaluate(
@@ -1782,7 +1782,7 @@ template sum_op<spin_handler> sum_op<spin_handler>::random(std::size_t nQubits,
                                                            std::size_t nTerms,
                                                            unsigned int seed);
 template csr_spmatrix sum_op<spin_handler>::to_sparse_matrix(
-    std::unordered_map<std::size_t, int> dimensions,
+    std::unordered_map<std::size_t, int64_t> dimensions,
     const std::unordered_map<std::string, std::complex<double>> &parameters,
     bool application_order) const;
 template std::vector<double>
@@ -1923,7 +1923,7 @@ std::tuple<std::vector<double>, std::size_t> sum_op<HandlerTy>::getDataTuple() c
 SPIN_OPS_BACKWARD_COMPATIBILITY_DEFINITION
 std::pair<std::vector<std::vector<bool>>, std::vector<std::complex<double>>>
 sum_op<HandlerTy>::get_raw_data() const {
-  std::unordered_map<std::size_t, int> dims;
+  std::unordered_map<std::size_t, int64_t> dims;
   auto degrees = this->degrees(
       false); // degrees in canonical order to match the evaluation
   auto evaluated = this->evaluate(
@@ -1971,7 +1971,7 @@ std::string sum_op<HandlerTy>::to_string(bool printCoeffs) const {
   // This function prints a string representing the operator sum that
   // includes the full representation for any degree in [0, max_degree),
   // padding identities if necessary (opposed to pauli_word).
-  std::unordered_map<std::size_t, int> dims;
+  std::unordered_map<std::size_t, int64_t> dims;
   auto degrees = this->degrees(
       false); // degrees in canonical order to match the evaluation
   auto evaluated = this->evaluate(
